% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weights.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate Precision Weights From Gamma Regression}
\usage{
calc_weights(data, gamma_reg_model)
}
\arguments{
\item{data}{A \code{data.frame} who's quantitative columns should be converted to
precision weights.}
}
\value{
The same \code{data.frame} but with all quantitative values replaced by
their precision weights
}
\description{
\code{calc_weights} returns a data frame with all quantitative values
replaced by their corresponding precision weights estimated from the gamma
regression.
}
\details{

}
\examples{
# Generate a design matrix for the data
design <- model.matrix(~ 0 + factor(rep(1:2, each = 3)))

# Set correct colnames, this is important for fit_gamma_weights
colnames(design) <- paste0("ng", c(50, 100))

# Normalize and log-transform the data
yeast <- prnn(yeast, "identifier")

# Fit the gamma regression model for the mean-variance trend
gamma_model <- fit_gamma_weights(yeast, design, "identifier")

# Generate the weights for the yeast data
calc_weights(yeast, gamma_model)
}
