% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_results}
\alias{extract_results}
\title{Title}
\usage{
extract_results(
  data,
  results,
  alpha = 0.05,
  abs_lfc = 1,
  pcor = stats::p.adjust.methods,
  id_col = "id"
)
}
\arguments{
\item{id_col}{}
}
\value{

}
\description{
Title
}
