% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma_regression.R
\name{plot_gamma_regression}
\alias{plot_gamma_regression}
\title{Function for plotting the mean-variance gamma regression}
\usage{
plot_gamma_regression(data, design, id_col = "id")
}
\arguments{
\item{data}{a \code{data.frame} to generate the mean-variance trends for. It
should contain columns with conditions named as the column names in
\code{design} (presumably with some suffix).}

\item{design}{a design or model matrix as produced my \code{model.matrix} with
column names corresponding to the different conditions.}

\item{id_col}{}
}
\value{

}
\description{
Function for plotting the mean-variance gamma regression
}
