% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma.R
\name{plot_ma}
\alias{plot_ma}
\title{Generate a MA-plot of the analysis}
\usage{
plot_ma(hits)
}
\arguments{
\item{hits}{a \code{tibble} produced by \code{\link[pair]{extract_results}}}
}
\value{
a \code{ggplot2} of the distribution of the hits
}
\description{
Generate a MA-plot of the analysis
}
\examples{

}
