% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norm_box.R
\name{plot_norm_box}
\alias{plot_norm_box}
\title{Generates a box-plot for \code{log2} transformed raw, tmm-, and psn- normalized
values.}
\usage{
plot_norm_box(
  data,
  id_col = "id",
  trim_M = 0.3,
  trim_A = 0.05,
  target = NULL,
  reference_sample = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to normalize}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.)}

\item{trim_M}{percent of fold-change values to trim}

\item{trim_A}{percent of means to trim}

\item{target}{Specify a subset of columns to normalize}

\item{reference_sample}{Specify a reference sample to normalize to in the
\code{\link[pair]{tmm}} method}
}
\value{
ggplot with samples on the x-axis and observed values on the y-axis,
different colors correspond to the raw or normalized data. Lines
corresponds to the global median across all samples.
}
\description{
This function can be used to produce a visual aid to for selection
normalization method. Ideally, after the data is normalized all
boxplots should have their median aligned close to the global trend line.
}
