% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prnn.R
\name{prnn}
\alias{prnn}
\title{Normalize data to a pseudo-reference}
\usage{
prnn(data, id_col = "id", log = TRUE, load_info = FALSE, target = NULL)
}
\arguments{
\item{data}{data.frame}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.)}

\item{log}{boolean variable indicating if the data should be log transformed
after normalization}

\item{load_info}{logical}

\item{target}{target columns to normalize, supports
\code{\link[tidyselect]{tidyselect-package}} syntax. By default, all numerical
columns will be used in the normalization if not specified.}
}
\value{
data.frame
}
\description{
Normalize data to a pseudo-reference
}
\examples{
yeast_prnn <- prnn(yeast, "identifier")
}
