% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_and_lfc.R
\name{run_limma_and_lfc}
\alias{run_limma_and_lfc}
\title{Title}
\usage{
run_limma_and_lfc(
  data,
  design,
  contrast_matrix,
  gamma_reg_model,
  id_col = "id"
)
}
\arguments{
\item{id_col}{}
}
\value{

}
\description{
Title
}
