% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Title}
\usage{
run_pipeline(
  data,
  design,
  contrast_matrix,
  imputations,
  workers,
  id_col = "id",
  plot_trend = F
)
}
\arguments{
\item{data}{a \code{data.frame} to generate the mean-variance trends for. It
should contain columns with conditions named as the column names in
\code{design} (presumably with some suffix).}

\item{design}{a design or model matrix as produced my \code{model.matrix} with
column names corresponding to the different conditions.}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.)}

\item{plot_trend}{}
}
\value{

}
\description{
Title
}
