% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmm.R
\name{tmm}
\alias{tmm}
\title{Normalization by Trimmed m Means}
\usage{
tmm(
  data,
  trim_M = 0.3,
  trim_A = 0.05,
  log = TRUE,
  load_info = FALSE,
  target = NULL,
  reference_sample = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to normalize}

\item{trim_M}{percent of fold-change values to trim}

\item{trim_A}{percent of means to trim}

\item{log}{Return log2 transformed values?}

\item{load_info}{Return loading info?}

\item{target}{target columns to normalize, supports
\code{\link[tidyselect]{tidyselect-package}} syntax. By default, all numerical
columns will be used in the normalization if not specified.}

\item{reference_sample}{Specify a reference sample to normalize to, if not
provided, the sample with the lowest coefficient of variation will be used}
}
\value{
data frame with normalized values
}
\description{
Normalization by Trimmed m Means
}
\examples{
tmm(yeast)
}
