% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outliers.R
\name{filter_outliers}
\alias{filter_outliers}
\title{Filter lowly abundant features}
\usage{
filter_outliers(data, target = NULL, percent = 1, k = 1.5)
}
\arguments{
\item{data}{data to filter from}

\item{target}{columns to base the filtering on}

\item{percent}{A feature gets filtered out if it is lowly abundant in \code{percent}
columns}

\item{k}{Parameter for the lower limit of Tukey's fence, any value bellow this
will be considered an outlier}
}
\value{
data with outliers removed
}
\description{
Filter lowly abundant features
}
