% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma_regression.R
\name{plot_gamma_regression}
\alias{plot_gamma_regression}
\title{Function for plotting the mean-variance gamma regression}
\usage{
plot_gamma_regression(data, design, id_col = "id")
}
\arguments{
\item{data}{The data to use for producing the plots}

\item{design}{A design matrix as produced by \code{\link[stats]{model.matrix}}.}

\item{id_col}{A character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.)}
}
\value{
a \code{ggplot2} object with the
}
\description{
Function for plotting the mean-variance gamma regression
}
