% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma_and_lfc.R
\name{run_limma_and_lfc}
\alias{run_limma_and_lfc}
\title{Run limma and calculate LFC}
\usage{
run_limma_and_lfc(
  data,
  design,
  contrast_matrix,
  gamma_reg_model = NULL,
  weights = NULL,
  id_col = "id"
)
}
\arguments{
\item{data}{a \code{data.frame} with the samples and feature ids}

\item{design}{a design or model matrix produced by
\code{\link[stats]{model.matrix}}}

\item{contrast_matrix}{a contrast matrix produced by
\code{\link[limma]{makeContrasts}}}

\item{gamma_reg_model}{the regression model produced by
\code{fit_gamma_weights} (see \code{\link[pair]{Mean-Variance_Gamma_Regressions}}) or any
\code{glm} with \code{formula} \code{sd ~ mean}}

\item{weights}{a matrix of precision weights}

\item{id_col}{a character for the name of the column containing the
name of the features in data (e.g., peptides, proteins, etc.)}
}
\value{

}
\description{
Function to calculate LFC, run limma using \code{\link[limma]{lmFit}}, \code{\link[limma]{contrasts.fit}},
and \code{\link[limma]{eBayes}} with the flag robust = TRUE.
If neither \code{weights} or \code{gamma_reg_model} is provided, then
\code{\link[limma]{lmFit}} will be run without precision weights.
If both are provided, then it will default to using the \code{gamma_reg_model}
model to produce the weights.
}
